refactor: move more functions to @vitest/mocker
