+++ packages/mocker/package.json
@@ -37,7 +37,11 @@
     "build": "rimraf dist && rollup -c",
     "dev": "rollup -c --watch"
   },
+  "dependencies": {
+    "magic-string": "^0.30.11"
+  },
   "devDependencies": {
+    "@vitest/utils": "workspace:*",
     "pathe": "^1.1.2"
   }
 }
