+++ test/core/package.json
@@ -14,6 +14,7 @@
   "devDependencies": {
     "@types/debug": "^4.1.12",
     "@vitest/expect": "workspace:*",
+    "@vitest/mocker": "workspace:*",
     "@vitest/runner": "workspace:*",
     "@vitest/test-dep1": "file:./deps/dep1",
     "@vitest/test-dep2": "file:./deps/dep2",
